修正pthread_attr_init()在MinGW(MSYS)和Cygwin环境下的一个问题; 修正jimi_getch()在MinGW(MSYS)和Cygwin环境下不会暂停的问题; 修正pop_list_verify()函数里, 当popmsg_list[i]的指针为NULL时导致的错误; 为VS 2013项目添加x64模式, 修正一些因为x64模式导致的问题.

+++ include/RingQueue/port.h
@@ -39,7 +39,7 @@
 #define jimi_b64(x)             (jimi_b32(x) | (jimi_b32(x) >> 32))
 
 #define jimi_next_power_of_2(x)     (jimi_b32((x) - 1) + 1)
-#define jimi_next_power_of_2_64(x)  (jimi_b64((x) - 1) + 1)
+#define jimi_next_power_of_2_64(x)  (jimi_b64((uint64_t)(x) - 1) + 1)
 
 #if defined(JIMI_SIZE_T_SIZEOF) && (JIMI_SIZE_T_SIZEOF == 8)
 #define JIMI_ROUND_TO_POW2(N)   jimi_next_power_of_2_64(N)
